feat: moved environment api to shared
